simplify the deregistration key and passport builder
deregistration doesn't need to have a proof since it will be added
in a block later, so we can make it smaller and more to the point.

passport builder needed to have a mean to add a new master key
but without owning that new master key's secret, now it is possible
since we can push a raw entry. Also it makes the flow easier to
manage having a mutable and immutable passport.